Fixed the debug, audio device, and naming (#15)

+++ src/pipeline.py
@@ -11,7 +11,7 @@ def run_pipeline():
     # Record audio
     recordings_dir = os.getenv("RECORDINGS_DIR", "/app/recordings")
     audio_recorder = AudioRecorder(recordings_dir, 44100)  # Will change this eventually to 32 kHz for energy saving
-    file_path= audio_recorder.record_audio(30)
+    file_path = audio_recorder.record_audio(30)
 
     gps_coords = {
         "lat": float(os.getenv("STATION_LATITUDE", "0.0")),
@@ -35,10 +35,11 @@ def run_pipeline():
         print(f"Inserted {len(detections)} detections.")
 
     finally:
-        # Always delete recording post processing
+        # Always delete recording post processing unless debugging
         if os.path.exists(file_path):
-            os.remove(file_path)
-            print("Deleted recording.")
+            if not int(os.getenv("DEBUG", "0")):
+                os.remove(file_path)
+                print("Deleted recording,", file_path)
 
     print("Pipeline run complete.")
